+++ modele/entitee/item/equippement/degainable/epee/epee.py
@@ -14,5 +14,5 @@ if TYPE_CHECKING:
 
 class Epee(Arme):
     """La classe des armes de type épée. Permettent de porter des coups semi-circulaires devant l'agissant."""
-    def __init__(self,labyrinthe:Labyrinthe,poids:float,frottements:float,element:Element,tranchant:float,portee:int,position:crt.Position=crt.POSITION_ABSENTE):
+    def __init__(self,labyrinthe:Labyrinthe,poids:float,frottements:float,element:Element,tranchant:float,portee:float,position:crt.Position=crt.POSITION_ABSENTE):
         Arme.__init__(self,labyrinthe,poids,frottements,element,tranchant,portee,position)
